refactor: move NaturalSelection
