+++ src/main/java/step2/DuplicateNumberPoolException.java
@@ -1,0 +1,4 @@
+package step2;
+
+class DuplicateNumberPoolException extends IllegalArgumentException {
+}
