+++ src/main/java/step2/IllegalCountException.java
@@ -1,0 +1,4 @@
+package step2;
+
+class IllegalCountException extends IllegalArgumentException {
+}
