+++ src/main/java/step2/NotEnoughNumberPoolSizeException.java
@@ -1,0 +1,4 @@
+package step2;
+
+class NotEnoughNumberPoolSizeException extends IllegalArgumentException {
+}
