+++ src/test/java/step2/NotEnoughMoneyException.java
@@ -1,0 +1,4 @@
+package step2;
+
+class NotEnoughMoneyException extends IllegalArgumentException {
+}
