+++ src/test/java/step2/NumberSelectionTest.java
@@ -43,33 +43,4 @@ public class NumberSelectionTest {
                 .isInstanceOf(IllegalCountException.class);
     }
 
-    public static class NaturalSelection {
-
-        public Set<Integer> select(List<Integer> numberPool, int count) {
-            Set<Integer> numberPoolSet = new HashSet<>(numberPool);
-            if (numberPoolSet.size() != numberPool.size()) {
-                throw new DuplicateNumberPoolException();
-            }
-            if (numberPoolSet.size() < count) {
-                throw new NotEnoughNumberPoolSizeException();
-            }
-            if (count < 1) {
-                throw new IllegalCountException();
-            }
-
-            List<Integer> modifiableNumberPool = new ArrayList<>(numberPool);
-            Collections.shuffle(modifiableNumberPool);
-            return new HashSet<>(modifiableNumberPool.subList(0, count));
-        }
-
-    }
-
-    private static class DuplicateNumberPoolException extends IllegalArgumentException {
-    }
-
-    private static class NotEnoughNumberPoolSizeException extends IllegalArgumentException {
-    }
-
-    private static class IllegalCountException extends IllegalArgumentException {
-    }
 }
